Add regression test for #173
