+++ tests/system/test_fixtures.py
@@ -1,0 +1,7 @@
+def test_new_fixture_name(scanomatic, browser):
+    """ Issue #173 """
+    browser.get(scanomatic + '/fixtures')
+    browser.find_element_by_id('add-fixture').click()
+    element = browser.find_element_by_id('new-fixture-name')
+    element.send_keys('Exterminate')
+    assert element.get_attribute('value') == "Exterminate"
